Move `AgentBatch` to `stateful`
