+++ packages/engine/lib/stateful/src/agent/arrow/mod.rs
@@ -1,4 +1,7 @@
 pub mod array;
+mod batch;
+
+pub use self::batch::AgentBatch;
 
 // TODO: this should be deleted, i.e. if this value is required use
 //      something like `get_hidden_column_name(PREVIOUS_INDEX_FIELD_NAME)`
