+++ packages/engine/src/datastore/batch/mod.rs
@@ -1,4 +1,4 @@
-pub mod agent;
+mod agent;
 pub mod boolean;
 pub mod context;
 pub mod dataset;
@@ -6,8 +6,9 @@ pub mod iterators;
 pub mod message;
 pub mod migration;
 
+pub use stateful::agent::AgentBatch;
+
 pub use self::{
-    agent::AgentBatch,
     context::{AgentIndex, ContextBatch},
     dataset::Dataset,
     message::MessageBatch,
